+++ spec/models/payroll_run_spec.rb
@@ -44,14 +44,30 @@ RSpec.describe PayrollRun, type: :model do
 
   describe ".create_with_claims!" do
     let(:claims) { Policies.all.map { |policy| create(:claim, :approved, policy: policy) } }
+    subject!(:payroll_run) { PayrollRun.create_with_claims!(claims, created_by: user) }
 
     it "creates a payroll run with payments and populates the award_amount" do
-      payroll_run = PayrollRun.create_with_claims!(claims, created_by: user)
-
       expect(payroll_run.reload.created_by.id).to eq(user.id)
       expect(payroll_run.claims).to match_array(claims)
       expect(claims[0].payment.award_amount).to eq(claims[0].award_amount)
       expect(claims[1].payment.award_amount).to eq(claims[1].award_amount)
+    end
+
+    context "with multiple claims with the same National Insurance number" do
+      let(:matching_claims) do
+        national_insurance_number = generate(:national_insurance_number)
+        [
+          create(:claim, :approved, policy: StudentLoans, national_insurance_number: national_insurance_number),
+          create(:claim, :approved, policy: MathsAndPhysics, national_insurance_number: national_insurance_number),
+        ]
+      end
+      let(:other_claim) { create(:claim, :approved) }
+      let(:claims) { matching_claims + [other_claim] }
+
+      it "groups them into a single payment and populates the award_amount" do
+        expect(payroll_run.payments.map(&:claims)).to match_array([match_array(matching_claims), [other_claim]])
+        expect(matching_claims[0].reload.payment.award_amount).to eq(matching_claims.sum(&:award_amount))
+      end
     end
   end